chore: move comparison to utils
