+++ tests/py/unit/test_tools_build_slab.py
@@ -8,6 +8,7 @@ from mat3ra.made.tools.build.slab import (
 from mat3ra.utils import assertion as assertion_utils
 
 from .fixtures import SI_SLAB_100
+from .utils import assert_two_entities_deep_almost_equal
 
 material = Material.create_default()
 
@@ -29,4 +30,4 @@ def test_build_slab():
     terminations = get_terminations(slab_config, build_parameters=params)
     termination = terminations[0]
     slab = create_slab(slab_config, termination, params)
-    assertion_utils.assert_deep_almost_equal(SI_SLAB_100, slab.to_json())
+    assert_two_entities_deep_almost_equal(slab, SI_SLAB_100)
